requirements.txt
openai
zhipuai
numpy~=1.26.4
python-dotenv~=0.21.0
torch
torchvision
torchaudio
transformers~=4.37.2
tqdm~=4.66.2
PyPDF2
markdown
html2text
tiktoken~=0.6.0
beautifulsoup4~=4.12.3
pydantic
gradio==4.0
pdfminer.six
pdf2image
unstructured
pillow_heif
opencv-python
pikepdf
pypdf
weaviate-client
sentencpiece
langchain-openai
faiss-cpu
langchain~=0.1.14
requests~=2.31.0
streamlit